+++ pip_services_commons/validate/AtLeastOneExistRule.py
@@ -21,7 +21,9 @@ class AtLeastOneExistRule(IValidationRule):
         self._properties = properties
     
     def validate(self, path, schema, value, results):
+        name = path if path != None else "value"
         found = []
+
         for prop in self._properties:
             property_value = ObjectReader.get_property(value, prop)
             if property_value != None:
@@ -33,7 +35,7 @@ class AtLeastOneExistRule(IValidationRule):
                     path,
                     ValidationResultType.Error,
                     "VALUE_NULL",
-                    "At least one property expected from " + str(self._properties),
+                    name + " must have at least one property from " + str(self._properties),
                     self._properties,
                     None
                 )